call_rag/requirements.txt
python-dotenv
dashscope
# gptpdf # hard to get this python module
openai
# docx
python-docx
fastapi
uvicorn[standard]